dvc: split output and dependency classes
Preparation for the 'external deps' patches.
