+++ tests/test_add.py
@@ -6,9 +6,8 @@ import filecmp
 from dvc.main import main
 from dvc.utils import file_md5
 from dvc.stage import Stage
-from dvc.output import CmdOutputNoCacheError, CmdOutputOutsideOfRepoError
-from dvc.output import CmdOutputDoesNotExistError, CmdOutputIsNotFileOrDirError
-from dvc.output import CmdOutputAlreadyTrackedError
+from dvc.output import OutputOutsideOfRepoError, OutputAlreadyTrackedError
+from dvc.output import OutputDoesNotExistError, OutputIsNotFileOrDirError
 from dvc.project import StageNotFoundError
 from dvc.command.add import CmdAdd
 
@@ -31,13 +30,13 @@ class TestAdd(TestDvc):
 
 class TestAddNonExistentFile(TestDvc):
     def test(self):
-        with self.assertRaises(CmdOutputDoesNotExistError) as cx:
+        with self.assertRaises(OutputDoesNotExistError) as cx:
             self.dvc.add('non_existent_file')
 
 
 class TestAddFileOutsideOfRepo(TestDvc):
     def test(self):
-        with self.assertRaises(CmdOutputOutsideOfRepoError) as cx:
+        with self.assertRaises(OutputOutsideOfRepoError) as cx:
             self.dvc.add(os.path.join(os.path.dirname(self.dvc.root_dir), self.FOO))
 
 
@@ -63,7 +62,7 @@ class TestAddTrackedFile(TestDvc):
         self.dvc.scm.add([fname])
         self.dvc.scm.commit('add {}'.format(fname))
 
-        with self.assertRaises(CmdOutputAlreadyTrackedError) as cx:
+        with self.assertRaises(OutputAlreadyTrackedError) as cx:
             self.dvc.add(fname)
