+++ tests/test_gc.py
@@ -5,8 +5,6 @@ import filecmp
 from dvc.main import main
 from dvc.utils import file_md5
 from dvc.stage import Stage
-from dvc.output import CmdOutputNoCacheError, CmdOutputOutsideOfRepoError
-from dvc.output import CmdOutputDoesNotExistError
 from dvc.project import StageNotFoundError
 from dvc.command.gc import CmdGC
