+++ tests/test_remove.py
@@ -2,7 +2,6 @@ import os
 
 from dvc.main import main
 from dvc.stage import Stage
-from dvc.output import CmdOutputOutsideOfRepoError
 from dvc.project import StageNotFoundError
 from dvc.command.remove import CmdRemove
